UI Fixes (call and pull request)
