--- Assets/PatchKit/UnityEditorExtension/Editor/UI/AccountScreen.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 62b7a723db864a979e343eac0eb25e6f
-timeCreated: 1537749058
