--- Assets/PatchKit/UnityEditorExtension/Editor/UI/NotLinkedAppScreen.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 947e477c4ed043d5b415b7960b26a834
-timeCreated: 1537760405
